Breaking out unit tests into separate files and upgrading them to include browser caching.
